Fix tests.
The primary issue was the LiveScript adding a header. I am guessing the
version of livescript got updated but the tests did not get updated to match.

Also the "fetch_state.js" file seemed to have Mac-style newlines (i.e. ^M) but
the expected result did not. Also it used actual tabs instead of spaces like
the other files. Just to ensure consistancy I changed everything on that file
to regular newlines (Unix style) an replaced the tabs with spaces.

Finally I added a bit of info to the README on how to run the tests. Since
it doesn't use a standard approach, it wasn't immediately obvious to me. The
documentation should hopeful help others that need to run the tests.