Fix error generated when @import used.
@import is often used in SASS/SCSS so variables declared in one file are
available in another. This library has code to correctly handled things so that
the @imported file used the asset pipeline. But it I am guessing some sort of
refectoring stopped it from working.

This commit adds the tests to demonstrate the problem (and to ensure it doesn't
break in the future). It also of course resolves the problem.